requirements.txt
streamlit==1.28.0
langchain==0.0.346
langchain-community==0.0.14
chromadb==0.4.15
sentence-transformers==2.2.2
pypdf==3.17.0
python-dotenv==1.0.0
tiktoken==0.5.1
google-generativeai==0.3.0
langchain-google-genai==0.0.2